Create splash screen Adjust names the activities
