+++ app/src/main/AndroidManifest.xml
@@ -9,15 +9,21 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".view.MainActivity">
+        <activity android:name=".view.SplashActivity"
+            android:theme="@style/Theme.AppCompat.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".view.AddProdutoActvity"/>
-        <activity android:name=".view.FinalizeActvity"/>
+
+        <activity android:name=".view.MainActivity"/>
+
+        <activity android:name=".view.AddProdutoActivity"/>
+
+        <activity android:name=".view.FinalizeActivity"/>
+
     </application>
 
 </manifest>
