+++ app/src/main/java/com/hgianastasio/kgarconbagual/view/MainActivity.kt
@@ -5,11 +5,9 @@ import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
-import android.widget.Toast
 import com.hgianastasio.kgarconbagual.R
 import com.hgianastasio.kgarconbagual.data.dao.ProdutoDAO
 import com.hgianastasio.kgarconbagual.data.database.DBHelper
-import com.hgianastasio.kgarconbagual.data.models.Produto
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
@@ -28,13 +26,13 @@ class MainActivity : AppCompatActivity() {
 
     override fun onOptionsItemSelected(item: MenuItem?): Boolean {
         when(item?.itemId){
-            R.id.addProduto_mi -> startActivityForResult(Intent(this,AddProdutoActvity::class.java), REQUSETCODE_ADD)
+            R.id.addProduto_mi -> startActivityForResult(Intent(this, AddProdutoActivity::class.java), REQUSETCODE_ADD)
         }
         return super.onOptionsItemSelected(item)
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if(requestCode== REQUSETCODE_ADD && resultCode == AddProdutoActvity.RESULT_OK){
+        if(requestCode== REQUSETCODE_ADD && resultCode == AddProdutoActivity.RESULT_OK){
             adapter = ProdutoAdapter(ProdutoDAO(DBHelper(this)).getAll(),this)
             lvProdutos.adapter = adapter
         }
@@ -52,7 +50,7 @@ class MainActivity : AppCompatActivity() {
         adapter?.qtdMap?.forEach {
             bundle.putInt(it.key.toString(),it.value)
         }
-        startActivity(Intent(this,FinalizeActvity::class.java).putExtras(bundle))
+        startActivity(Intent(this, FinalizeActivity::class.java).putExtras(bundle))
     }
 
     companion object {
